Move other common parts to shared folder (Cell primitives, StateInit, etc) Move out Everscale dependencies from CommonTON (e.g. Wallet-related InitData, createSigningMessage and so on) Remove left Everscale dependencies from TheOpenNetwork (replaced with CommonTON) Cleanup unused includes
